Refactored token endpoint helpers and added support for the two remaining flows: client credentials and resource owner password credentials.
