+++ tests/test_tandem_10_oauth2_token_exchange.py
@@ -401,7 +401,7 @@ class TestEndpoint(object):
         scope, audience and subject_token_type works.
         """
         endp = self.server.get_endpoint("token")
-        conf = endp.helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
+        conf = endp.grant_type_helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
         conf["policy"][""]["kwargs"] = {}
         conf["policy"][""]["kwargs"]["audience"] = ["https://example.com"]
         conf["policy"][""]["kwargs"]["resource"] = ["https://example.com"]
@@ -436,7 +436,7 @@ class TestEndpoint(object):
         grant_types_supported (that are set in its helper attribute).
         """
         endpoint = self.server.get_endpoint("token")
-        del endpoint.helper["urn:ietf:params:oauth:grant-type:token-exchange"]
+        del endpoint.grant_type_helper["urn:ietf:params:oauth:grant-type:token-exchange"]
 
         resp, _state, _scope = self.process_setup()
 
@@ -452,10 +452,7 @@ class TestEndpoint(object):
         _te_request, _te_resp = self.do_query("token_exchange", "token", req_args, _state)
 
         assert _te_resp["error"] == "invalid_request"
-        assert (
-                _te_resp["error_description"]
-                == "Unsupported grant_type: urn:ietf:params:oauth:grant-type:token-exchange"
-        )
+        assert _te_resp["error_description"] == "Do not know how to handle this type of request"
 
     def test_wrong_resource(self):
         """
@@ -463,7 +460,7 @@ class TestEndpoint(object):
         """
         endpoint = self.server.get_endpoint("token")
 
-        conf = endpoint.helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
+        conf = endpoint.grant_type_helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
         conf["policy"][""]["kwargs"] = {}
         conf["policy"][""]["kwargs"]["resource"] = ["https://example.com"]
 
@@ -512,7 +509,7 @@ class TestEndpoint(object):
         Test that requesting a token for an unknown audience fails.
         """
         endpoint = self.server.get_endpoint("token")
-        conf = endpoint.helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
+        conf = endpoint.grant_type_helper["urn:ietf:params:oauth:grant-type:token-exchange"].config
         conf["policy"][""]["kwargs"] = {}
         conf["policy"][""]["kwargs"]["audience"] = ["https://example.com"]
